Fix build errors.
	- OpenGL stuff that got hosed via the merge.
	- package rename stuff.

Also updated tags file ;).

+++ core/src/main/java/com/spidey01/sxe/core/OpenGLES20.java
@@ -369,7 +369,7 @@ public interface OpenGLES20 {
     void glDisable(int cap);
     void glDisableVertexAttribArray(int index);
     void glDrawElements(int mode, int count, int type, ByteBuffer indices);
-    void glDrawElements(int mode, int count, int type, long offset);
+    void glDrawElements(int mode, int count, int type, int offset);
     void glDrawArrays(int mode, int first, int count);
 
     void glEnable(int cap);
